\name{box.scale}
% [1] "al"             "box.scale"      "calout.detect"  "ckesd"         
% [5] "gesdri"         "hamp.scale.3"   "hamp.scale.4"   "hampor"        
% [9] "hampoutinds"    "lamtab"         "logit"          "prompt.default"
%[13] "rouor"          "rououtinds"     "shorth"         "shorth.scale"  
%[17] "skesd"          "tukeyor"        "tukeyorinds"   
%
\alias{box.scale}
\alias{hamp.scale.3}
\alias{hamp.scale.4}
\alias{shorth.scale}
\title{ calibrated scaling inlier multiplier radius for various outlier detection approaches }
\description{ calibrated scaling inlier multiplier radius for various outlier detection approaches }
\usage{
box.scale(n, alpha=0.05)
}
\arguments{
  \item{n}{ n}
  \item{alpha}{ alpha}
}

\author{Vince Carey <stvjc@channing.harvard.edu>}

\examples{
box.scale(20)
}
\keyword{ models }

